[MPN] Add Branch Argument
It has started to seem like it would be better if we passed the branch
of the build in. We seem to want to know it everywhere. So I added it as
an argument to the rprel inputs. Anywhere that calls rprel would need to
be updated to add this argument to the rprel call.

But it means moving forward there is a direct connection between the
build and what branch it was created from.

I bumped the major version since changes were breaking.

I also bumped the elixir version in the mix file to match the exenv file
which is 1.3.3