[MPN] Bump timeout on file uploads
We tried adding `rprel release` to ag, but it failed because of a
timeout while uploading the tarball. It seems like since we are upload
files as big as (in this case) 350 MB, we should be giving it a lot more
time. So I bumped it from 10,000 milliseconds (10 seconds) to 100,000
milliseconds (100 seconds).

I also placed that value in config so its easier to change in the future.